Test 3D tile added
I wanted to rename its components but I have no idea if that would break things, I suspect this is kindof a bad direction to go it for things like web builds
I'll likely just make traditional environment assets. if it's too much trouble to implement this. don't do it